Fixes card styling (#72)
* Fix: Moving away the Avatar Image to a reusable component that can be centered, the Avatar Component is very specific to the lists.

* Fix: Fixed the user account view layout as well.

* Remove unused var
